requirements.txt
# TCA IRR Backend Requirements
# Core web framework and database
fastapi==0.115.4
uvicorn[standard]==0.32.0
asyncpg==0.30.0

# Authentication and security
python-jwt==4.1.0
PyJWT==2.10.1
bcrypt==4.2.1
python-multipart==0.0.18

# Data validation and serialization
pydantic[email]==2.10.3

# Environment and configuration
python-dotenv==1.0.1

# Async utilities
asyncio-extras==1.3.2

# Development dependencies
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1